+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- The `accessibilityIdentifier` can now be set on `AttributedLabel`.
+
 ### Removed
 
 ### Changed